ARC-401 Add PHP Apache images

+++ builds/run.sh
@@ -12,6 +12,9 @@ fi
 
 BUILD_SCRIPT=$1
 
+DEV_MODE="${DEV_MODE:false}"
+export DEV_MODE;
+
 CI_DOCKER_REGISTRY="${CI_DOCKER_REGISTRY:-registry.hub.docker.com}"
 export CI_DOCKER_REGISTRY;
 
@@ -53,22 +56,39 @@ if [ -z ${CI_DOCKER_NAMESPACE+x} ] || [ -z "${CI_DOCKER_NAMESPACE}" ]; then
     exit 1
 fi
 
-#If we have Docker Username and password login to docker
-if [ ! -z ${CI_DOCKER_USERNAME+x} ] && [ "${CI_DOCKER_USERNAME}" != "" ] && [ ! -z ${CI_DOCKER_TOKEN+x} ] && [ "${CI_DOCKER_TOKEN}" != "" ]; then
-    echo "logging into registry"
-    echo "${CI_DOCKER_TOKEN}" | docker login --username=$CI_DOCKER_USERNAME --password-stdin $CI_DOCKER_REGISTRY
-    export CI_ACTION_PUSH_IMAGES=true
+# Check if in Dev Mode
+if [ "$DEV_MODE" = true ]; then
+  echo "Dev MODE Enabled Prevent Pushing to docker hub"
+else
+  #If we have Docker Username and password login to docker
+  if [ "${CI_DOCKER_REGISTRY}" == "" ]; then
+    echo "AWS Login"
+    
+  else
+    if [ ! -z ${CI_DOCKER_USERNAME+x} ] && [ "${CI_DOCKER_USERNAME}" != "" ] && [ ! -z ${CI_DOCKER_TOKEN+x} ] && [ "${CI_DOCKER_TOKEN}" != "" ]; then
+        echo "logging into registry"
+        echo "${CI_DOCKER_TOKEN}" | docker login --username=$CI_DOCKER_USERNAME --password-stdin $CI_DOCKER_REGISTRY
+        export CI_ACTION_PUSH_IMAGES=true
+    fi
+  fi
 fi
 
-docker buildx create --name base_img_builder --use
+DOCKER_BUILD_COMMAND="docker buildx build ${CI_BUILD_ARGS} --platform=${CI_BUILD_PLATFORMS}";
+if [ "$DEV_MODE" = true ]; then
+  DOCKER_BUILD_COMMAND="docker build ${CI_BUILD_ARGS}";
+else
+  docker buildx create --name base_img_builder --use
+fi
 
 echo ">>> Running build script '${BUILD_SCRIPT}.sh'";
 ${SCRIPTPATH}/${BUILD_SCRIPT}.sh
 ret_code=$?
 
-if [ "$BUILD_DELETE_BUILDX_BUILDER_AFTER_BUILD" = true ]; then
-    echo ">>> Removeing base_img_builder"
-    docker buildx rm base_img_builder
+if [ "$DEV_MODE" = false ]; then
+  if [ "$BUILD_DELETE_BUILDX_BUILDER_AFTER_BUILD" = true ]; then
+      echo ">>> Removeing base_img_builder"
+      docker buildx rm base_img_builder
+  fi
 fi
 
 exit $ret_code
